third_party/WebKit/LayoutTests/platform/win7/virtual/outofblink-cors-http/http/tests/fetch/serviceworker/thorough/cors-base-https-other-https-expected.txt

This is a testharness.js-based test.
PASS Startup 
PASS cors 
PASS executeTest-0 
PASS executeTest-1 
PASS executeTest-2 
PASS executeTest-3 
PASS executeTest-4 
FAIL executeTest-5 assert_equals: https://localhost:8443/serviceworker/resources/fetch-access-control.php?mode=cors&credentials=same-origin&method=GET&ACAOrigin=* should be resolved expected "resolved" but got "rejected"
FAIL executeTest-6 assert_equals: https://localhost:8443/serviceworker/resources/fetch-access-control.php?mode=cors&credentials=same-origin&method=GET&ACAOrigin=https://127.0.0.1:8443 should be resolved expected "resolved" but got "rejected"
PASS executeTest-7 
PASS executeTest-8 
FAIL executeTest-9 assert_equals: https://localhost:8443/serviceworker/resources/fetch-access-control.php?mode=cors&credentials=same-origin&method=GET&ACAOrigin=*&ACEHeaders=X-ServiceWorker-ServerHeader should be resolved expected "resolved" but got "rejected"
FAIL executeTest-10 assert_equals: https://localhost:8443/serviceworker/resources/fetch-access-control.php?mode=cors&credentials=same-origin&method=GET&ACAOrigin=https://127.0.0.1:8443&ACEHeaders=X-ServiceWorker-ServerHeader should be resolved expected "resolved" but got "rejected"
FAIL executeTest-11 assert_equals: https://localhost:8443/serviceworker/resources/fetch-access-control.php?mode=cors&credentials=same-origin&method=GET&ACAOrigin=*&ACEHeaders=Content-Length, X-ServiceWorker-ServerHeader should be resolved expected "resolved" but got "rejected"
FAIL executeTest-12 assert_equals: https://localhost:8443/serviceworker/resources/fetch-access-control.php?mode=cors&credentials=same-origin&method=GET&ACAOrigin=https://127.0.0.1:8443&ACEHeaders=Content-Length, X-ServiceWorker-ServerHeader should be resolved expected "resolved" but got "rejected"
PASS executeTest-13 
FAIL executeTest-14 assert_equals: https://localhost:8443/serviceworker/resources/fetch-access-control.php?mode=cors&credentials=same-origin&method=POST&ACAOrigin=* should be resolved expected "resolved" but got "rejected"
FAIL executeTest-15 assert_equals: https://localhost:8443/serviceworker/resources/fetch-access-control.php?mode=cors&credentials=same-origin&method=POST&ACAOrigin=https://127.0.0.1:8443 should be resolved expected "resolved" but got "rejected"
PASS executeTest-16 
PASS executeTest-17 
FAIL executeTest-18 assert_equals: https://localhost:8443/serviceworker/resources/fetch-access-control.php?mode=cors&credentials=same-origin&method=POST&ACAOrigin=*&ACEHeaders=X-ServiceWorker-ServerHeader should be resolved expected "resolved" but got "rejected"
FAIL executeTest-19 assert_equals: https://localhost:8443/serviceworker/resources/fetch-access-control.php?mode=cors&credentials=same-origin&method=POST&ACAOrigin=https://127.0.0.1:8443&ACEHeaders=X-ServiceWorker-ServerHeader should be resolved expected "resolved" but got "rejected"
FAIL executeTest-20 assert_equals: https://localhost:8443/serviceworker/resources/fetch-access-control.php?mode=cors&credentials=same-origin&method=POST&ACAOrigin=*&ACEHeaders=Content-Length, X-ServiceWorker-ServerHeader should be resolved expected "resolved" but got "rejected"
FAIL executeTest-21 assert_equals: https://localhost:8443/serviceworker/resources/fetch-access-control.php?mode=cors&credentials=same-origin&method=POST&ACAOrigin=https://127.0.0.1:8443&ACEHeaders=Content-Length, X-ServiceWorker-ServerHeader should be resolved expected "resolved" but got "rejected"
Harness: the test ran to completion.
